Rename file to fix typo
