+++ src/vs/workbench/contrib/notebook/browser/chatEdit/notebookSynchronizer.ts
